beam-starter-python/requirements.txt
apache-beam==2.46.0
beam-mysql-connector